完成階段性修改
1. 將 header 拆出成獨立元件。
2. 加強 alert 功能並調整結構。
3. 待完成的功能增加施工中訊息。
4. 調整連結、按鈕的 hover 行為。

5. 改寫 loading 功能：
(1) 將區域 loading 改寫成 composable。
(2) 將全域 loading 改寫成 pinia 以便全域管理狀態。

6. 調整 Banner 及輪播相關：
(1) 檔案 BannerSlide 更名為 Banner。
(2) Banner 使用 swiper 輪播。
(3) Albums 輪播增加顯示頁碼。
(4) Swiper style 改在 main.js 匯入。因為在任一隻 Vue 檔案匯入都會變成全域，那不如直接於 main.js 匯入。
