+++ src/assets/scss/base/_variables.scss
@@ -41,9 +41,9 @@ $font-family-sans-serif:
 // 顏色
 $primary: #030303 !default;
 $secondary: #fbff22 !default;
-// $success: #198754 !default;
-// $info: #0dcaf0 !default;
-// $warning: #ffc107 !default;
+$success: #198754 !default;
+$info: #0dcaf0 !default;
+$warning: #ffc107 !default;
 $danger: #dc3545 !default;
 // $light: #f8f9fa !default;
 // $dark: #212529 !default;
@@ -53,9 +53,9 @@ $theme-colors: (
   "primary": $primary,
   "secondary": $secondary,
   "notice": #c96464,
-  // "success": $success,
-  // "info": $info,
-  // "warning": $warning,
+  "success": $success,
+  "info": $info,
+  "warning": $warning,
   "danger": $danger,
   // "light": $light,
    //  "dark": $dark,
