+++ src/components/eventDetail/TicketInfo.vue
@@ -29,7 +29,7 @@
             <button
               type="button"
               class="px-3 px-md-6 py-2 btn btn-outline-primary font-noto-serif-tc"
-              @click.prevent
+              @click.prevent="purchase(ticket)"
             >
               <span class="d-none d-md-inline">立即</span>購票
             </button>
@@ -41,17 +41,23 @@
 </template>
 
 <script>
-// TODO 待檢查優化
+import { mapActions } from "pinia";
+import { useLoadingStore } from "@/stores";
+import { ConfirmAlert } from "@/helpers";
 import { EventTagRecord } from "@/services/data/event";
 // eslint-disable-next-line no-unused-vars
 import { TicketRecord, TicketModel, TicketService } from "@/services/data/ticket";
 
 export default {
+  /**
+   * @property {EventTagRecord} event 活動資料。
+   */
   props: {
-    event: EventTagRecord,
+    event: {
+      type: EventTagRecord,
+      required: true,
+    },
   },
-
-  inject: ["loading"],
 
   /**
    * @returns {{ tickets: TicketRecord[] }}
@@ -63,7 +69,8 @@ export default {
   },
 
   created() {
-    this.loading.open();
+    console.log(`## [TicketInfo] this ==>`, this);
+    this.openLoading();
 
     // 取得票券資訊
     TicketService.fetchEventTickets(this.event.id)
@@ -71,8 +78,28 @@ export default {
         this.tickets = new TicketModel(res).datas;
       })
       .finally(() => {
-        this.loading.close();
+        this.closeLoading();
       });
+  },
+
+  methods: {
+    /** loading 功能 */
+    ...mapActions(useLoadingStore, {
+      openLoading: "open",
+      closeLoading: "close",
+    }),
+
+    /**
+     * 購買票券。
+     * @param {TicketRecord} ticket 被點擊的票券資料。
+     */
+    // eslint-disable-next-line no-unused-vars
+    purchase(ticket) {
+      // console.log(`ticket ===>`, ticket);
+      // console.log(`event ===>`, this.event);
+      // TODO 購票功能
+      ConfirmAlert.alertWarningDetail("施工中...", "購票功能尚未開放，敬請期待！");
+    },
   },
 };
 </script>
