+++ src/composables/index.js
@@ -2,6 +2,10 @@
 import { usePagination, PaginationParam } from "./usePagination";
 export { usePagination, PaginationParam };
 
+// loading
+import { useAreaLoading } from "./loading/useAreaLoading";
+export { useAreaLoading };
+
 // features
 import { useEventTag } from "./data/useEventTag";
 export { useEventTag };
