+++ src/helpers/index.js
@@ -6,10 +6,13 @@ import { DateStyles } from "./date/dateStyles";
 import { DateUtils } from "./date/dateUtils";
 import { DataUtils } from "./dataUtils";
 import { HistoryStateHelper } from "./HistoryStateHelper";
-import { FullLoadingHelper } from "./loadingHelper";
-import { AreaLoadingHelper } from "./loadingHelper";
+import { LoadingSettings } from "./loadingHelper";
 import { VeeValidateHelper } from "./veeValidateHelper";
-import { AlertOptions, BaseAlert, ConfirmAlert, TimerAlert } from "./sweetAlertHelper";
+
+// alert
+import { AlertOptions } from "./alert/alertOptions";
+import { BasicAlert } from "./alert/basicAlert";
+import { ConfirmAlert, TimerAlert } from "./alert/alertHelper";
 
 // event
 import { KeyboardDisableHelper } from "./event/keyboardHelper";
@@ -23,12 +26,11 @@ export {
   DateUtils,
   DataUtils,
   HistoryStateHelper,
-  FullLoadingHelper,
-  AreaLoadingHelper,
+  LoadingSettings,
   VeeValidateHelper,
   KeyboardDisableHelper,
   AlertOptions,
-  BaseAlert,
+  BasicAlert,
   ConfirmAlert,
   TimerAlert,
 };
