+++ src/stores/index.js
@@ -1,0 +1,3 @@
+import { useLoadingStore } from "./loadingStore";
+
+export { useLoadingStore };
